Use host IP addresses only in the case of RegisterProxy
Because a host can have several interfaces, this change eliminates
associating any host IP address with a service statically but instead
only relies on dynamically discovering IP addresses of the interface
associated with the incoming query.

In case of RegisterProxy() where a list of IPs can be specified during
service registration, the provided list is associated with host IP
addresses to be used in the response.